Update to confix.txt comments.
Comment variables (comment text in brackets) edited so that all comments follow a '0 to 1' scheme. (0 first as it is the default for a majority of these options).

'(optional)' moved inside brackets on all comments where it is present (this was previously inconsistent).
Various edits for general comment readability.

The term 'file name' was maintained, although 'filename' would be more appropriate. I will review this when the config.txt gains more instances of 'file name' in the future.